Refactor "Id" to be a Long type instead of Integer. Issue #17

+++ Extensions/ContentQuery/src/main/java/com/bingzer/android/dbv/content/queries/InsertImpl.java
@@ -33,7 +33,7 @@ public class InsertImpl extends com.bingzer.android.dbv.internal.queries.InsertI
     }
 
     @Override
-    public Integer query(){
+    public Long query(){
         return UriUtils.parseIdFromUri(uri);
     }
