+++ Extensions/ContentQuery/src/main/java/com/bingzer/android/dbv/content/resolvers/Resolver.java
@@ -37,14 +37,14 @@ public final class Resolver extends BaseResolver implements IResolver {
     }
 
     @Override
-    public int selectId(String whereClause, Object... args) {
-        int id = -1;
+    public long selectId(String whereClause, Object... args) {
+        long id = -1;
         Cursor cursor = null;
         try{
             cursor = select(whereClause, args).query();
             if(cursor.moveToNext()){
                 int index = cursor.getColumnIndex(getPrimaryKeyColumn());
-                id = cursor.getInt(index);
+                id = cursor.getLong(index);
             }
         }
         finally {
@@ -85,7 +85,7 @@ public final class Resolver extends BaseResolver implements IResolver {
     ///////////////////////////////////////////////////////////////////
 
     @Override
-    public Select select(int id) {
+    public Select select(long id) {
         return select(generateParamId(id));
     }
 
@@ -105,7 +105,7 @@ public final class Resolver extends BaseResolver implements IResolver {
     }
 
     @Override
-    public Select select(int... ids) {
+    public Select select(long... ids) {
         if(ids != null && ids.length > 0){
             StringBuilder whereClause = new StringBuilder();
             whereClause.append(getPrimaryKeyColumn()).append(" ");
